+++ src/main/java/com/lowdragmc/mbd2/integration/jei/MBDRecipeTypeCategory.java
@@ -70,7 +70,7 @@ public class MBDRecipeTypeCategory extends ModularUIRecipeCategory<MBDRecipeType
                 registration.addRecipes(MBDRecipeTypeCategory.TYPES.apply(recipeType),
                         Minecraft.getInstance().getConnection().getRecipeManager().getAllRecipesFor(recipeType)
                                 .stream()
-                                .filter(recipe -> !recipe.isFuel)
+                                .filter(recipe -> !recipe.isFuel && !recipe.isXEIHidden)
                                 .map(RecipeWrapper::new)
                                 .collect(Collectors.toList()));
             }