+++ CHANGELOG.md
@@ -3,4 +3,5 @@
 ## v1.0.31
 * Fixed World Preview renderer doesn't work correctly with forge model
 * Added PNC Pressure Condition
+* Improve XEI recipe viewer control. (added control for proxy recipe, and specific recipe)